% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{cluster_analysis}
\alias{cluster_analysis}
\title{k-means Clustering Analysis of Principal Components}
\usage{
cluster_analysis(
  pca,
  num_clusters = NA,
  max_clusters = 10,
  x = "PC1",
  y = "PC2",
  nstart = 10
)
}
\arguments{
\item{pca}{matrix of principal component values for each observation}

\item{num_clusters}{number of clusters to use for analysis; if left blank, the function will use the optimal number of clusters determined by the elbow method}

\item{max_clusters}{maximum number of clusters to analyze for determining the optimal number of clusters}

\item{x}{the principal component to be plotted on the x axis of the scatterplot}

\item{y}{the principal component to be plotted on the y axis of the scatterplot}

\item{nstart}{number of times to run k-means clustering}
}
\description{
This function performs k-means clustering on PCA results to group similar observations. It also determines the optimal number of clusters using the elbow method and generates plots.
}
\examples{
p <- pca(iris)
cluster_analysis(p$pca, num_clusters = NA, max_clusters = 10, x = "PC1", y = "PC2", nstart = 10)
}
