% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Principal Component Analysis}
\usage{
pca(data, eigen_opt = NA, threshold = 85)
}
\arguments{
\item{data}{matrix of data set to be analyzed}

\item{eigen_opt}{determines number of principal components to calculate; if left blank, the function will use the optimal number of components determined by the "threshold" argument}

\item{threshold}{percent of variance the minimum number of principal components must explain}
}
\description{
This allows for the exploration of the Principle component analysis of mass spectometry proteomics data
}
\examples{
pca(iris, eigen_opt = NA, threshold = 85)
}
\keyword{analysis}
\keyword{component}
\keyword{eigen}
\keyword{eigenvalue}
\keyword{eigenvalues}
\keyword{eigenvector}
\keyword{eigenvectors}
\keyword{pca}
\keyword{principal}
