% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCA_interface}
\alias{PCA_interface}
\title{Principal Component Analysis}
\usage{
PCA_interface(data_mat, num_components)
}
\arguments{
\item{data_mat}{numeric data matrix [n x p]}

\item{num_components}{int desired number of components}
}
\value{
Lower dimension representation of input matrix  [n x num_components]
}
\description{
Principal Component Analysis
}
