% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree Plot}
\usage{
scree_plot(eigen_values, threshold = 85, y_axis = "Eigenvalue")
}
\arguments{
\item{eigen_values}{Matrix of eigenvalues}

\item{threshold}{percent of variance the minimum number of principal components must explain}

\item{y_axis}{y axis for the scree plot; use "Eigenvalue" for eigenvalues to be plotted and "Percentage" for percentage of explained variance}
}
\description{
Generates a scree plot, determines the optimal number of principal components, and calculates the total proportion of variance explained
}
\examples{
p <- pca(iris)
scree_plot(p$eigen_values, threshold = 85, y_axis = "Eigenvalue")
}
\keyword{analysis}
\keyword{component}
\keyword{eigen}
\keyword{eigenvalue}
\keyword{eigenvalues}
\keyword{pca}
\keyword{principal}
